[phase-12-bugfix] fix user-visible P1 bugs and cosmetic cleanups
P1 (user-visible):
- B1: Profile 'Overall Accuracy' showed 0% always — operator-precedence
  bug ((data ?? 0.0) * 100) parsed as (data ?? (0.0 * 100)). Now
  (((data ?? 0.0).toDouble() * 100).toStringAsFixed(0)) matches the
  Today's Accuracy formatter.
- B2: Mastery retry dialog displayed placeholder total instead of real
  correctAnswers / totalInteractionCount. Adds public getters on
  LessonViewModel and uses them in the dialog subtitle.
- B4: StudyStatsProvider._randomSuffix used DateTime.now().millisecond
  for every one of the 6 chars, giving 6 identical chars and ~1/1000
  entropy. Now uses Random.nextInt(36).
- B5: getWeakWords was an empty-list stub. Now parses MistakeEntry log,
  aggregates by wordId (or grammarPointId fallback), sorts by
  mistakeCount desc, and resolves displayTerm/translation via the
  vocab / grammar maps.
- B7: interactionCorrectAnswerLabel(ListenOnly) leaked audioAsset path
  into the mistake log when no transcript was set. Now returns null
  (ListenOnly is listen-and-continue, no correct answer).

Cosmetic / analyzer:
- D1: statistics.dart used hard-coded VarnamalaTheme.textHint which
  breaks dark mode. Now uses textHintColor(context).
- D2: drop stale 'TODO: removed when AutoRoute is introduced' in
  extensions.dart and the duplicated route extension.
- D3-D6: analysis_options.yaml now ignores invalid_annotation_target
  (freezed @JsonKey on factory params), unnecessary_null_comparison
  and unnecessary_non_null_assertion in drift-generated code.
- D7: capture Navigator + Theme before the showDialog await in
  new_lesson_screen.dart to silence use_build_context_synchronously.

Tests: +17 regression tests (115/115 green, was 98/98).

+++ lib/views/profile/widgets/learning_stats.dart
@@ -260,7 +260,8 @@ class _OverallStatsGrid extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final totalMinutes = (data['totalMinutes'] as num?)?.toInt() ?? 0;
-    final accuracy = ((data['accuracy'] as num?) ?? 0.0 * 100).toStringAsFixed(0);
+    final accuracy = (((data['accuracy'] as num?) ?? 0.0).toDouble() * 100)
+        .toStringAsFixed(0);
     final totalLessons = (data['totalLessons'] as num?)?.toInt() ?? 0;
     final totalReviews = (data['totalReviews'] as num?)?.toInt() ?? 0;
